+++ src/cjs/Line.d.ts
@@ -17,15 +17,17 @@
  * @modified 2022-10-17 Adding these methods from the `PathSegment` interface: getStartPoint, getEndPoint, revert.
  * @modified 2023-09-25 Changed param type of `intersection()` from Line to VertTuple.
  * @modified 2025-04-15 Class `Line` now implements interface `Intersectable`.
+ * @modified 2025-04-16 Class `Line` now implements interface `IBounded`.
  * @version  2.4.0
  *
  * @file Line
  * @public
  **/
+import { Bounds } from "./Bounds";
 import { Vector } from "./Vector";
 import { VertTuple } from "./VertTuple";
 import { Vertex } from "./Vertex";
-import { Intersectable, PathSegment, SVGSerializable } from "./interfaces";
+import { IBounded, Intersectable, PathSegment, SVGSerializable } from "./interfaces";
 /**
  * @classdesc A line consists of two vertices a and b.<br>
  * <br>
@@ -34,7 +36,7 @@ import { Intersectable, PathSegment, SVGSerializable } from "./interfaces";
  *
  * @requires Vertex
  */
-export declare class Line extends VertTuple<Line> implements Intersectable, PathSegment, SVGSerializable {
+export declare class Line extends VertTuple<Line> implements IBounded, Intersectable, PathSegment, SVGSerializable {
     /**
      * Required to generate proper CSS classes and other class related IDs.
      **/
@@ -58,6 +60,15 @@ export declare class Line extends VertTuple<Line> implements Intersectable, Path
      * @memberof Line
      **/
     intersection(line: VertTuple<any>): Vertex | null;
+    /**
+     * Get the bounding box (bounds) of this Line.
+     *
+     * @method getBounds
+     * @instance
+     * @memberof Line
+     * @return {Bounds} The rectangular bounds of this Line.
+     **/
+    getBounds(): Bounds;
     /**
      * Get the start point of this path segment.
      *